Adapt the CBS export job to check the NVR before starting the build
This patch modifies the behavior of the export script.
The NVR is first checked between HEAD and HEAD^1.

If the NVR has not changed (that means it is likely to be
a DLRN change to follow the upstream stable branch changes) then
the job skips the CBS build and exit 0 to allow the change
to land in the repository. No CBS build is done there.

If the NVR has changed the job starts a non-scratch build
and exit 0 (thus the change lands in the repository) only if
the CBS build succeed.

This patch is part of the story
https://trello.com/c/URAtrhLU/86-automate-stable-packages-releases

Change-Id: I93f49cf6a75dc110f6c87cd173e5984af6ed113c